requirements.txt
numpy
Theano==0.8
EasyDict
Pillow
pyyaml
sklearn
pycuda
